requirements.txt
numpy
tqdm
scipy
scikit-learn
yacs
albumentations
torch
torchvision
timm
Pillow
matplotlib
pydantic
flask
gunicorn
python-json-logger
